update spec, refactor homepage a bit
