+++ clients/spas/app/pages/HomePage.vue
