+++ clients/spas/app/router.ts
@@ -1,5 +1,5 @@
 import { createWebHistory, createRouter } from "vue-router";
-import HomePage from "./views/HomePage.vue";
+import HomePage from "./pages/HomePage.vue";
 
 const routes = [{ path: "/", component: HomePage }];
